Address compiler errors

+++ src/dijkstras.hpp
@@ -6,7 +6,7 @@
 class dijkstras {
     public:
     dijkstras();
-        std::vector<float> dijkstra_path(std::vector<std::vector<int>> graph, int start);
+        std::vector<float> dijkstra_path(std::vector<std::vector<double> > graph, int start);
     private:
         int minDistance(std::vector<float> distances, std::vector<int> shortest_path);
 };
